Finished all basic user and product business logical and sql implementation

+++ src/main/java/com/geespring/marketapi/service/ProductService.java
@@ -1,6 +1,7 @@
 package com.geespring.marketapi.service;
 
 import com.geespring.marketapi.model.Product;
+import com.geespring.marketapi.model.User;
 import com.geespring.marketapi.repository.ProductRepository;
 import com.geespring.marketapi.util.exceptions.DataNotFoundException;
 import com.google.common.collect.ImmutableList;
@@ -8,6 +9,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
@@ -24,7 +26,7 @@ public class ProductService {
      * Fetches all the products from the database
      * @return list of every single {@link Product}
      */
-    public List<Product> getAllProducts() {
+    public List<Product> getAll() {
         return ImmutableList.copyOf(repository.findAll().iterator());
     }
 
@@ -46,10 +48,36 @@ public class ProductService {
     }
 
     /**
+     *
+     * @param id
+     * @param name
+     * @param description
+     * @param price
+     */
+    @Transactional
+    public void update(final Long id, String name, String description, Float price) {
+        Product existing = repository.findById(id).orElseThrow(DataNotFoundException::new);
+
+        if (name != null && name.length() > 0 && !name.equalsIgnoreCase(existing.getName())) {
+            existing.setName(name);
+        }
+
+        if (description != null && description.length() > 0 && !description.equalsIgnoreCase(existing.getDescription())) {
+            existing.setDescription(description);
+        }
+
+        if (price != null && !existing.getPrice().equals(price)) {
+            existing.setPrice(price);
+        }
+    }
+
+    /**
      * Deletes the {@link Product} with the specified id.
      * @param id
      */
     public void delete(final Long id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else throw new IllegalStateException("Product with id " + id + " does not exist.");
     }
 }
